chore(raft): drop raft implementation for now
The complexity of the project bumps significantly with the
addition of distributed computing challenges. To move faster
on the core features that do not require cluster of Bouines,
this commit drops the support in server.go of raft. This will
be re-introduced later on once this project is fullfilling all
single node expectations.