refact(naming): rename bouineX to bouine
This follows the temporary drop of all distributed system
references. For the sake of simplicity, for the moment single
node efficiency and compliance is going to remain the main
target until reaching a fully stable and performant state.

Then RAFT implementation through the Hashicorp library is
going to be the next step.