test(rename): rename smoke tests into e2e
Those tests are indeed not smoke tests but rather end-to-end (e2e)
tests. To adopt a way faster and way simpler manner to run
integration tests, this bouine rework will focus on using
the power of the stdlib with httptest and leverage pure Go
integration tests that can be run in parallel every now and
then without relying on external dependencies and Docker.